improvements to numeric least sq. fitting.
Much hinges on the method used to determine whether an a-vector is feasible. In my tests, in practise feasibilityViaCDFSumNegative() works best, but I don't understand why. The one that seems theoretically best according to the paper is feasibilityViaPDF(), but this produces poor results. I would like to understand this better.

Three methods for feasibility:
- feasibilityViaCDFSumNegative (must be equal to 0)
- feasibilityViaPDF (must be equal to 0)
- CDFSlopeNumericMin (must be non-negative)